requirements.txt
# Web 框架
Flask==2.3.3
Flask-CORS==4.0.0

# Markdown 处理
Markdown==3.5.1
Pygments==2.17.2  # 代码高亮支持

# 异步任务
celery==5.3.4
redis==5.0.1

# HTTP 请求
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 语音识别
faster-whisper==0.10.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

# 讲话人识别
pyannote.audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0

# NLP 处理
jieba==0.42.1
scikit-learn==1.3.2

# 模板引擎
Jinja2==3.1.2

# 进度条
tqdm==4.66.1

# 配置文件
PyYAML==6.0.1
python-dotenv==1.0.0  # 环境变量管理

# 日志
loguru==0.7.2

# AI SDK（可选）
openai>=2.0.0  # OpenAI/DeepSeek/豆包/通义千问（OpenAI 兼容接口）
anthropic==0.7.0  # Claude
dashscope==1.14.1  # 阿里云通义千问（语音识别）

# 开发工具
pytest==7.4.3
black==23.12.0
flake8==6.1.0
